requirements-dev.txt
when-changed==0.3.0
python-dotenv==0.6.0
pytest==3.0
pytest-django==3.0
pytest-env==0.6
django-debug-toolbar==1.6
